footswitch: Move GPIO switch logic to a base class
This moves the GPIO handling and detection of long vs short presses
to a base class that can be reused for other switches
