+++ pistomp/footswitch.py
@@ -17,14 +17,13 @@ import logging
 import RPi.GPIO as GPIO
 from rtmidi.midiconstants import CONTROL_CHANGE
 
-import pistomp.controller as controller
-import time
+import pistomp.gpioswitch as gpioswitch
 import queue
 
-class Footswitch(controller.Controller):
+class Footswitch(gpioswitch.GpioSwitch):
 
     def __init__(self, id, fs_pin, led_pin, midi_CC, midi_channel, midiout, refresh_callback):
-        super(Footswitch, self).__init__(midi_channel, midi_CC)
+        super(Footswitch, self).__init__(fs_pin, midi_channel, midi_CC)
         self.id = id
         self.display_label = None
         self.enabled = False
@@ -35,29 +34,16 @@ class Footswitch(controller.Controller):
         self.relay_list = []
         self.preset_callback = None
         self.lcd_color = None
-        self.cur_tstamp = None
-        self.events = queue.Queue()
-
-        # this value (in seconds) chosen to be just greater than the event_detect bouncetime (in milliseconds)
-        self.relay_poll_interval = 0.26
-        self.relay_poll_intervals = 2
-
-        # Long press threshold in seconds
-        self.long_press_threshold = 0.5
-
-        GPIO.setup(fs_pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
-        GPIO.add_event_detect(fs_pin, GPIO.FALLING, callback=self._pressed, bouncetime=250)
 
         if led_pin is not None:
             GPIO.setup(led_pin, GPIO.OUT)
             self._set_led(GPIO.LOW)
 
-    def __del__(self):
-        GPIO.remove_event_detect(self.fs_pin)
-
+    # Should this be in Controller ?
     def set_midi_CC(self, midi_CC):
         self.midi_CC = midi_CC
 
+    # Should this be in Controller ?
     def set_midi_channel(self, midi_channel):
         self.midi_channel = midi_channel
 
@@ -72,49 +58,7 @@ class Footswitch(controller.Controller):
     def set_lcd_color(self, color):
         self.lcd_color = color
 
-    def _pressed(self, gpio):
-        # This is run from a separate thread, timestamp pressed and queue an event
-        #
-        # I considered using a dual edge callback and handle the timestamp here
-        # to queue long/short press events, but in practice, I noticed dual edge
-        # is rather unreliable with such a long debounce, we often don't get the
-        # rising edge callback at all. So let's just timestamp and we'll handle
-        # everything from the poller thread
-        #
-        self.events.put(time.monotonic())
-
-
-    def poll(self):
-        # Grab press event if any
-        if not self.events.empty():
-            new_tstamp = self.events.get_nowait()
-        else:
-            new_tstamp = None
-
-        # If we were a already pressed and waiting for a release, drop it, it's easier
-        # that way and we should be polling fast enough for this not to matter.
-        # Otherwise record it
-        if self.cur_tstamp is None:
-            self.cur_tstamp = new_tstamp
-
-        # Are we waiting for release ?
-        if self.cur_tstamp is None:
-            return
-
-        time_pressed = time.monotonic() - self.cur_tstamp
-
-        # If it's a long press, process as soon as we reach the threshold, otherwise
-        # check the GPIO input
-        if time_pressed > self.long_press_threshold:
-            short = False
-        elif GPIO.input(self.fs_pin):
-            short = True
-        else:
-            return
-        self.cur_tstamp = None
-
-        logging.debug("Footswitch %d %s press" % (self.fs_pin, "short" if short else "long"))
-
+    def pressed(self, short):
         # If a footswitch can be mapped to control a relay, preset, MIDI or all 3
         #
         # The footswitch will only "toggle" if it's associated with a relay
